[#89] refactor: Chat 패키지 구조 변경
