+++ src/main/java/chat/teco/tecochat/chat/presentation/chat/api/request/AskRequest.java
@@ -1,0 +1,6 @@
+package chat.teco.tecochat.chat.presentation.chat.api.request;
+
+public record AskRequest(
+        String message
+) {
+}
