+++ src/main/java/chat/teco/tecochat/chat/presentation/chat/api/request/CreateChatRequest.java
@@ -1,4 +1,4 @@
-package chat.teco.tecochat.chat.presentation.chat.request;
+package chat.teco.tecochat.chat.presentation.chat.api.request;
 
 public record CreateChatRequest(
         String message
