+++ src/main/java/chat/teco/tecochat/chat/presentation/chat/api/request/UpdateChatTitleRequest.java
@@ -1,4 +1,4 @@
-package chat.teco.tecochat.chat.presentation.chat.request;
+package chat.teco.tecochat.chat.presentation.chat.api.request;
 
 public record UpdateChatTitleRequest(
         String title
