+++ src/main/java/chat/teco/tecochat/chat/presentation/chat/api/response/CopyChatResponse.java
@@ -1,4 +1,4 @@
-package chat.teco.tecochat.chat.presentation.chat.response;
+package chat.teco.tecochat.chat.presentation.chat.api.response;
 
 public record CopyChatResponse(
         Long copiedChatId
