+++ src/main/java/chat/teco/tecochat/chat/presentation/chat/api/response/CreateChatResponse.java
@@ -1,4 +1,4 @@
-package chat.teco.tecochat.chat.presentation.chat.response;
+package chat.teco.tecochat.chat.presentation.chat.api.response;
 
 public record CreateChatResponse(
         Long chatId,
