+++ src/main/java/chat/teco/tecochat/common/config/AuthenticationConfig.java
@@ -1,6 +1,6 @@
 package chat.teco.tecochat.common.config;
 
-import chat.teco.tecochat.auth.argumentresolver.AuthArgumentResolver;
+import chat.teco.tecochat.auth.presentation.argumentresolver.AuthArgumentResolver;
 import java.util.List;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
